+++ frontend/src/components/common/Footer.tsx
