+++ frontend/src/components/header/DetailedCategory.module.css
