+++ frontend/src/components/header/DetailedCategory.tsx
