+++ frontend/src/components/header/SearchItemList.tsx
