+++ frontend/src/index.tsx
@@ -7,30 +7,30 @@ import {createBrowserRouter, RouterProvider} from 'react-router-dom';
 import {QueryClient, QueryClientProvider,} from '@tanstack/react-query';
 import ProductPage from './pages/category/ProductPage';
 import Root from './pages/home/Root';
-import Login from './pages/auth/Login';
-import Signup from './pages/auth/Signup';
-import IdFind from './pages/auth/IdFind';
+import Login from './pages/auth/login/Login';
+import Signup from './pages/auth/signup/Signup';
+import IdFind from './pages/auth/find/IdFind';
 import store from './redux/store/store';
-import Agreement from './pages/auth/Agreement';
-import Exhibitions from './pages/category/Exhibitions';
+import Agreement from './pages/auth/signup/Agreement';
+import Exhibitions from './pages/category/exhibitions/Exhibitions';
 import MoreInformation from "./pages/moreInformation/MoreInformation";
-import DeliveryInquiry from "./pages/userService/DeliveryInquiry";
+import DeliveryInquiry from "./pages/myPage/mySubPage/DeliveryInquiry";
 import Information from "./pages/moreInformation/Information";
 import SellerInformation from "./pages/moreInformation/SellerInformation";
 import Cart from "./pages/userService/Cart";
 import SearchPage from "./pages/category/SearchPage";
-import MyPage from "./pages/userService/MyPage";
-import TimeDeal from "./pages/category/TimeDeal";
+import MyPage from "./pages/myPage/MyPage";
+import TimeDeal from "./pages/category/timedeal/TimeDeal";
 import ProtectedRoute from "./utils/route/ProtectedRoute";
 import RestrictedRoute from "./utils/route/RestrictedRoute";
 import Review from "./pages/moreInformation/review";
 import Payment from "./pages/userService/Payment";
-import UserInfo from "./pages/userService/mySubPage/UserInfo";
-import OrderManagement from "./pages/userService/mySubPage/OrderManagement";
-import AddressManagement from "./pages/userService/mySubPage/AddressManagement";
-import PointsAndCoupons from "./pages/userService/mySubPage/PointsAndCoupons";
-import InquiryHistory from "./pages/userService/mySubPage/InquiryHistory";
-import PasswordFind from "./pages/auth/PasswordFind";
+import UserInfo from "./pages/myPage/mySubPage/UserInfo";
+import OrderManagement from "./pages/myPage/mySubPage/OrderManagement";
+import AddressManagement from "./pages/myPage/mySubPage/AddressManagement";
+import PointsAndCoupons from "./pages/myPage/mySubPage/PointsAndCoupons";
+import InquiryHistory from "./pages/myPage/mySubPage/InquiryHistory";
+import PasswordFind from "./pages/auth/find/PasswordFind";
 import {ReactQueryDevtools} from "@tanstack/react-query-devtools";
 
 const router = createBrowserRouter([
