+++ frontend/src/layout/Header/SearchInput.tsx
@@ -9,12 +9,13 @@ import {FaMagnifyingGlass} from "../../icons/Icons";
 // TODO: 최근검색어 구현하기
 const SearchInput = () => {
     const dispatch = useDispatch<AppDispatch>();
-    const inputValue = useSelector((state: RootState) => state.search.input);
     const location = useLocation();
     const navigate = useNavigate();
-    const [isInputClicked, setIsInputClicked] = useState<boolean>(false);
-    const recentSearches: string[] = ['검색어1', '검색어2', '검색어3']; // 예시 데이터
     const inputContainerRef = useRef<HTMLDivElement>(null);
+    const inputValue = useSelector((state: RootState) => state.search.input);
+    const [isInputClicked, setIsInputClicked] = useState<boolean>(false);
+    // TODO: 로컬스토리지에 최근 검색어 저장하기 (최대 10개 최신순으로)
+    const recentSearches: string[] = ['검색어1', '검색어2', '검색어3'];
 
     useEffect(() => {
         dispatch(setSearchInput(''));
@@ -31,11 +32,6 @@ const SearchInput = () => {
         dispatch(setSearchInput(e.target.value));
     };
 
-    const handleSearchClick = () => {
-        navigateToSearchResults(inputValue);
-        setIsInputClicked(false); // 검색 버튼 클릭 시 닫기
-    };
-
     const handleInputToggle = () => {
         setIsInputClicked(prevState => !prevState);
     };
@@ -44,6 +40,15 @@ const SearchInput = () => {
         if (inputContainerRef.current && !inputContainerRef.current.contains(e.relatedTarget as Node)) {
             setIsInputClicked(false);
         }
+    };
+
+    const navigateToSearchResults = (query: string) => {
+        navigate(`/search?query=${encodeURIComponent(query)}`);
+    };
+
+    const handleSearchClick = () => {
+        navigateToSearchResults(inputValue);
+        setIsInputClicked(false);
     };
 
     const handleOutsideClick = (e: MouseEvent) => {
@@ -62,10 +67,6 @@ const SearchInput = () => {
             navigateToSearchResults(inputValue);
             setIsInputClicked(false);
         }
-    };
-
-    const navigateToSearchResults = (query: string) => {
-        navigate(`/search?query=${encodeURIComponent(query)}`);
     };
 
     // 빈 문자열을 제외한 검색어 필터링
@@ -100,9 +101,9 @@ const SearchInput = () => {
                                 <button className={styles.close__btn} onClick={handleInputToggle}>전체삭제</button>
                             </div>
                             <div className={styles.historyItem__container}>
-
-                                {filteredSearches.map((search) => (
-                                    <li key={search} className={styles.historyItem}
+                                {filteredSearches.map((search, _index) => (
+                                    <li key={search}
+                                        className={styles.historyItem}
                                         onClick={() => handleSearchItemClick(search)}>
                                         <FaMagnifyingGlass/>
                                         {search}
