+++ frontend/src/pages/myPage/mySubPage/DeliveryInquiry.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styles from './DeliveryInquiry.module.css';
-import ImageUpload from "../../utils/adminsMaster/ImageUpload";
+import ImageUpload from "../../../utils/adminsMaster/ImageUpload";
 
 const DeliveryInquiry: React.FC = () => {
     return (
