+++ frontend/src/pages/myPage/mySubPage/InquiryHistory.tsx
