+++ frontend/src/pages/myPage/mySubPage/UserInfo.tsx
@@ -4,7 +4,7 @@ import axiosInstance from "../../../utils/error/axiosInstance";
 import { handleAxiosError } from "../../../utils/error/errorHandler";
 import { AxiosError } from "axios";
 import LoadingSpinner from "../../../utils/reactQuery/LoadingSpinner";
-import Footer from "../../../components/home/Footer";
+import Footer from "../../../components/common/Footer";
 
 interface Authority {
     authority: string;
